requirements.txt
gunicorn
flask
africastalking
pymongo
python-dotenv
flask
flask-pymongo
pymongo
dnspython
pycountry
phonenumbers
bcrypt